Stopped passing addresses of unaligned fields in packed structures.
GCC 9 in Fedora 30 now issues warnings for this which are
converted to errors, breaking the build. These unaligned
references are technically incorrect usage that GCC had always
allowed and we were able to get away with on X86 because it mostly
doesn't care about alignment.

.../blockMap.c
.../fixedLayout.c
.../header.c
.../recoveryJournal.c
.../slabDepot.c
.../vdo.c
.../volumeGeometry.c
.../parseUtils.c
Stopped passing addresses of unaligned fields to buffer decoding
  functions, replacing them with local variables that are safely
  copied back into the fields through the structure reference.

Author: Michael Sclafani <sclafani@redhat.com>